chore: split selectNextPart into its own file
