+++ packages/job-worker/src/playout/activePlaylistActions.ts
@@ -5,7 +5,8 @@ import _ = require('underscore')
 import { JobContext } from '../jobs'
 import { logger } from '../logging'
 import { CacheForPlayout, getOrderedSegmentsAndPartsFromPlayoutCache, getSelectedPartInstancesFromCache } from './cache'
-import { resetRundownPlaylist, selectNextPart } from './lib'
+import { resetRundownPlaylist } from './lib'
+import { selectNextPart } from './selectNextPart'
 import { setNextPart } from './setNext'
 import { updateStudioTimeline, updateTimeline } from './timeline/generate'
 import { getCurrentTime } from '../lib'
