+++ packages/job-worker/src/playout/take.ts
@@ -5,7 +5,8 @@ import { UserError, UserErrorMessage } from '@sofie-automation/corelib/dist/erro
 import { logger } from '../logging'
 import { JobContext, ProcessedShowStyleCompound } from '../jobs'
 import { CacheForPlayout, getOrderedSegmentsAndPartsFromPlayoutCache, getSelectedPartInstancesFromCache } from './cache'
-import { isTooCloseToAutonext, selectNextPart } from './lib'
+import { isTooCloseToAutonext } from './lib'
+import { selectNextPart } from './selectNextPart'
 import { setNextPart } from './setNext'
 import { getCurrentTime } from '../lib'
 import { PartEndState, VTContent } from '@sofie-automation/blueprints-integration'
